requirements.txt
requests~=2.32.0
beautifulsoup4~=4.12.3
pillow~=10.3.0